Successfully Done
K means clustering is one of the most popular clustering algorithms and usually the first thing
practitioners apply when solving clustering tasks to get an idea of the structure of the dataset.
The goal of K means is to group data points into distinct non-overlapping subgroups. One of
the major application of K means clustering is segmentation of customers to get a better
understanding of them which in turn could be used to increase the revenue of the company